Correct renames. Better CheckPhysicalCameraProperties.cs test
