+++ com.unity.cinemachine.hdrptests/Tests/Runtime/com.unity.cinemachine.hdrp.runtime.tests.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "com.unity.cinemachine.hdrpruntime.tests",
+    "name": "com.unity.cinemachine.hdrp.runtime.tests",
     "rootNamespace": "",
     "references": [
         "com.unity.cinemachine",
@@ -8,7 +8,8 @@
         "UnityEngine.TestRunner",
         "UnityEditor.TestRunner",
         "Unity.Splines",
-        "Unity.Mathematics"
+        "Unity.Mathematics",
+        "com.unity.cinemachine.hdrp.tests"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
